// src/types/Walks/Walk.ts
export default interface Walk {
    aboutTheAuthor: string;
    author: string;
    city: string;
    content1: string;
    content2: string;
    content3: string;
    coverImg: string;
    description: string;
    facebookLink: string;
    featuredWalk: boolean;
    iframeLink: string;
    iframeTitle: string;
    img1: string;
    img2: string;
    img3: string;
    instagramLink: string;
    lat: string;
    length: string;
    lng: string;
    mapImg: string;
    startingPoint: string;
    twitterLink: string;
    walk: string;
    websiteLink: string;
    __v: number;
    _id: string;
}